fix: Hide preview on the very small screesns
Also a bit refactors the layout code
