+++ tests/screenshots/tests-picker_ui_snap.lua---debug---file_info_panel_bottom
