+++ tests/screenshots/tests-picker_ui_snap.lua---default---cursor_second_item_bottom
