--- tests/screenshots/tests-picker_ui_snap.lua---narrow---cursor_second_item
@@ -1,51 +1,0 @@
---|---------|---------|---------|---------|---------|---------|---------|
-01|                                                                      
-02|~                                                                     
-03|~       ╭ README.md ───────────────────────────────────────────╮      
-04|~       │# Fixture                                             │      
-05|~       │                                                      │      
-06|~       │                                                      │      
-07|~       │                                                      │      
-08|~       │                                                      │      
-09|~       │                                                      │      
-10|~       │                                                      │      
-11|~       │                                                      │      
-12|~       │                                                      │      
-13|~       ├ FFFiles ─────────────────────────────────────────────┤      
-14|~       │  intro.md docs                                       │      
-15|~       │  guide.md docs                                       │      
-16|~       │  contributing.md docs                                │      
-17|~       │  changelog.md docs                                   │      
-18|~       │  README.md                                           │      
-19|~       ├──────────────────────────────────────────────────────┤      
-20|~       │>                                                 32  │      
-21|~       ╰──────────────────────────────────────────────────────╯      
-22|~                                                                     
-23|[No Name]                                           0,1            All
-24|-- INSERT --                                        1,3           All 
-
---|---------|---------|---------|---------|---------|---------|---------|
-01|0000000000000000000000000000000000000000000000000000000000000000000000
-02|1111111111111111111111111111111111111111111111111111111111111111111111
-03|1111111123333333333322222222222222222222222222222222222222222222111111
-04|1111111124444444445555555555555555555555555555555555555555555552111111
-05|1111111126666666666666666666666666666666666666666666666666666662111111
-06|1111111126666666666666666666666666666666666666666666666666666662111111
-07|1111111126666666666666666666666666666666666666666666666666666662111111
-08|1111111126666666666666666666666666666666666666666666666666666662111111
-09|1111111126666666666666666666666666666666666666666666666666666662111111
-10|1111111126666666666666666666666666666666666666666666666666666662111111
-11|1111111126666666666666666666666666666666666666666666666666666662111111
-12|1111111126666666666666666666666666666666666666666666666666666662111111
-13|1111111123333333332222222222222222222222222222222222222222222222111111
-14|1111111126655555555577775555555555555555555555555555555555555552111111
-15|1111111126655555555577775555555555555555555555555555555555555552111111
-16|1111111126655555555555555557777555555555555555555555555555555552111111
-17|1111111126655555555555557777555555555555555555555555555555555552111111
-18|1111111128899999999999999999999999999999999999999999999999999992111111
-19|1111111122222222222222222222222222222222222222222222222222222222111111
-20|1111111125555555555555555555555555555555555555555555555555566552111111
-21|1111111122222222222222222222222222222222222222222222222222222222111111
-22|1111111111111111111111111111111111111111111111111111111111111111111111
-23|::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-24|;;;;;;;;;;;;<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
