+++ tests/screenshots/tests-picker_ui_snap.lua---narrow---cursor_second_item_bottom
