+++ tests/screenshots/tests-picker_ui_snap.lua---narrow---cursor_second_item_top
@@ -30,22 +30,22 @@
 03|1111111123333333332222222222222222222222222222222222222222222222111111
 04|1111111124444444444444444444444444444444444444444444444444455442111111
 05|1111111122222222222222222222222222222222222222222222222222222222111111
-06|1111111126677777777777777777777777777777777777777777777777777772111111
-07|1111111125544444444444448888444444444444444444444444444444444442111111
-08|1111111125544444444444444448888444444444444444444444444444444442111111
-09|1111111125544444444488884444444444444444444444444444444444444442111111
-10|1111111125544444444488884444444444444444444444444444444444444442111111
-11|1111111125544444444444888844444444444444444444444444444444444442111111
-12|1111111125544444444444448888444444444444444444444444444444444442111111
-13|1111111125544444448884444444444444444444444444444444444444444442111111
-14|1111111125544444448884444444444444444444444444444444444444444442111111
-15|1111111125544444444448884444444444444444444444444444444444444442111111
-16|1111111125544444444488844444444444444444444444444444444444444442111111
-17|1111111125544444448884444444444444444444444444444444444444444442111111
-18|1111111125544444444888444444444444444444444444444444444444444442111111
+06|1111111125544444444444444444444444444444444444444444444444444442111111
+07|1111111126677777777777778888777777777777777777777777777777777772111111
+08|1111111125544444444444444449999444444444444444444444444444444442111111
+09|1111111125544444444499994444444444444444444444444444444444444442111111
+10|1111111125544444444499994444444444444444444444444444444444444442111111
+11|1111111125544444444444999944444444444444444444444444444444444442111111
+12|1111111125544444444444449999444444444444444444444444444444444442111111
+13|1111111125544444449994444444444444444444444444444444444444444442111111
+14|1111111125544444449994444444444444444444444444444444444444444442111111
+15|1111111125544444444449994444444444444444444444444444444444444442111111
+16|1111111125544444444499944444444444444444444444444444444444444442111111
+17|1111111125544444449994444444444444444444444444444444444444444442111111
+18|1111111125544444444999444444444444444444444444444444444444444442111111
 19|1111111122222222222222222222222222222222222222222222222222222222111111
 20|1111111111111111111111111111111111111111111111111111111111111111111111
 21|1111111111111111111111111111111111111111111111111111111111111111111111
 22|1111111111111111111111111111111111111111111111111111111111111111111111
-23|9999999999999999999999999999999999999999999999999999999999999999999999
-24|::::::::::::;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+23|::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+24|;;;;;;;;;;;;<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
