+++ tests/screenshots/tests-picker_ui_snap.lua---narrow---no_results_top
@@ -2,8 +2,8 @@
 01|                                                                      
 02|~                                                                     
 03|~       ╭ FFFiles ─────────────────────────────────────────────╮      
-04|~       │                                                      │      
-05|~       │                                                      │      
+04|~       │> zzzzzzzzz                                     0/32  │      
+05|~       ├──────────────────────────────────────────────────────┤      
 06|~       │                                                      │      
 07|~       │                                                      │      
 08|~       │                                                      │      
@@ -17,9 +17,9 @@
 16|~       │                                                      │      
 17|~       │                                                      │      
 18|~       │                                                      │      
-19|~       ├──────────────────────────────────────────────────────┤      
-20|~       │> zzzzzzzzz                                     0/32  │      
-21|~       ╰──────────────────────────────────────────────────────╯      
+19|~       ╰──────────────────────────────────────────────────────╯      
+20|~                                                                     
+21|~                                                                     
 22|~                                                                     
 23|[No Name]                                           0,1            All
 24|-- INSERT --                                        1,12          All 
@@ -28,24 +28,24 @@
 01|0000000000000000000000000000000000000000000000000000000000000000000000
 02|1111111111111111111111111111111111111111111111111111111111111111111111
 03|1111111123333333332222222222222222222222222222222222222222222222111111
-04|1111111124455555555555555555555555555555555555555555555555555552111111
-05|1111111124455555555555555555555555555555555555555555555555555552111111
-06|1111111124455555555555555555555555555555555555555555555555555552111111
-07|1111111124455555555555555555555555555555555555555555555555555552111111
-08|1111111124455555555555555555555555555555555555555555555555555552111111
-09|1111111124455555555555555555555555555555555555555555555555555552111111
-10|1111111124455555555555555555555555555555555555555555555555555552111111
-11|1111111124455555555555555555555555555555555555555555555555555552111111
-12|1111111124455555555555555555555555555555555555555555555555555552111111
-13|1111111124455555555555555555555555555555555555555555555555555552111111
-14|1111111124455555555555555555555555555555555555555555555555555552111111
-15|1111111124455555555555555555555555555555555555555555555555555552111111
-16|1111111124455555555555555555555555555555555555555555555555555552111111
-17|1111111124455555555555555555555555555555555555555555555555555552111111
-18|1111111124455555555555555555555555555555555555555555555555555552111111
+04|1111111124444444444444444444444444444444444444444444444445555442111111
+05|1111111122222222222222222222222222222222222222222222222222222222111111
+06|1111111125544444444444444444444444444444444444444444444444444442111111
+07|1111111125555555555555555555555555555555555555555555555555555552111111
+08|1111111125555555555555555555555555555555555555555555555555555552111111
+09|1111111125555555555555555555555555555555555555555555555555555552111111
+10|1111111125555555555555555555555555555555555555555555555555555552111111
+11|1111111125555555555555555555555555555555555555555555555555555552111111
+12|1111111125555555555555555555555555555555555555555555555555555552111111
+13|1111111125555555555555555555555555555555555555555555555555555552111111
+14|1111111125555555555555555555555555555555555555555555555555555552111111
+15|1111111125555555555555555555555555555555555555555555555555555552111111
+16|1111111125555555555555555555555555555555555555555555555555555552111111
+17|1111111125555555555555555555555555555555555555555555555555555552111111
+18|1111111125555555555555555555555555555555555555555555555555555552111111
 19|1111111122222222222222222222222222222222222222222222222222222222111111
-20|1111111125555555555555555555555555555555555555555555555554444552111111
-21|1111111122222222222222222222222222222222222222222222222222222222111111
+20|1111111111111111111111111111111111111111111111111111111111111111111111
+21|1111111111111111111111111111111111111111111111111111111111111111111111
 22|1111111111111111111111111111111111111111111111111111111111111111111111
 23|6666666666666666666666666666666666666666666666666666666666666666666666
 24|7777777777778888888888888888888888888888888888888888888888888888888888
