+++ tests/screenshots/tests-picker_ui_snap.lua---narrow---query_main_bottom
@@ -1,17 +1,17 @@
 --|---------|---------|---------|---------|---------|---------|---------|
 01|                                                                      
 02|~                                                                     
-03|~       ╭ src/main.rs ─────────────────────────────────────────╮      
-04|~       │fn main() {}                                          │      
-05|~       │                                                      │      
-06|~       │                                                      │      
-07|~       │                                                      │      
-08|~       │                                                      │      
-09|~       │                                                      │      
-10|~       │                                                      │      
-11|~       │                                                      │      
-12|~       │                                                      │      
-13|~       ├ FFFiles ─────────────────────────────────────────────┤      
+03|~       ╭ FFFiles ─────────────────────────────────────────────╮      
+04|~       │  api.rs src                                          │      
+05|~       │  license.md docs                                     │      
+06|~       │  regression.rs tests                                 │      
+07|~       │  menu.tsx src/components                             │      
+08|~       │  changelog.md docs                                   │      
+09|~       │  contributing.md docs                                │      
+10|~       │  integration.rs tests                                │      
+11|~       │  input.tsx src/components                            │      
+12|~       │  intro.md docs                                       │      
+13|~       │  main_test.rs tests                                  │      
 14|~       │  main_utils.rs src                                   │      
 15|~       │  main_runner.rs src                                  │      
 16|~       │  main_loop.rs src                                    │      
@@ -27,24 +27,24 @@
 --|---------|---------|---------|---------|---------|---------|---------|
 01|0000000000000000000000000000000000000000000000000000000000000000000000
 02|1111111111111111111111111111111111111111111111111111111111111111111111
-03|1111111123333333333333222222222222222222222222222222222222222222111111
-04|1111111124444444444444444444444444444444444444444444444444444442111111
-05|1111111125555555555555555555555555555555555555555555555555555552111111
-06|1111111125555555555555555555555555555555555555555555555555555552111111
-07|1111111125555555555555555555555555555555555555555555555555555552111111
-08|1111111125555555555555555555555555555555555555555555555555555552111111
-09|1111111125555555555555555555555555555555555555555555555555555552111111
-10|1111111125555555555555555555555555555555555555555555555555555552111111
-11|1111111125555555555555555555555555555555555555555555555555555552111111
-12|1111111125555555555555555555555555555555555555555555555555555552111111
-13|1111111123333333332222222222222222222222222222222222222222222222111111
-14|1111111125566664444444444777444444444444444444444444444444444442111111
-15|1111111125566664444444444477744444444444444444444444444444444442111111
-16|1111111125566664444444447774444444444444444444444444444444444442111111
-17|1111111125566664444444444477744444444444444444444444444444444442111111
-18|1111111128866669999:::999999999999999999999999999999999999999992111111
+03|1111111123333333332222222222222222222222222222222222222222222222111111
+04|1111111124455555556665555555555555555555555555555555555555555552111111
+05|1111111124455555555555666655555555555555555555555555555555555552111111
+06|1111111124455555555555555666665555555555555555555555555555555552111111
+07|1111111124455555555566666666666666555555555555555555555555555552111111
+08|1111111124455555555555556666555555555555555555555555555555555552111111
+09|1111111124455555555555555556666555555555555555555555555555555552111111
+10|1111111124455555555555555566666555555555555555555555555555555552111111
+11|1111111124455555555556666666666666655555555555555555555555555552111111
+12|1111111124455555555566665555555555555555555555555555555555555552111111
+13|1111111124477775555555556666655555555555555555555555555555555552111111
+14|1111111124477775555555555666555555555555555555555555555555555552111111
+15|1111111124477775555555555566655555555555555555555555555555555552111111
+16|1111111124477775555555556665555555555555555555555555555555555552111111
+17|1111111124477775555555555566655555555555555555555555555555555552111111
+18|1111111128877779999:::999999999999999999999999999999999999999992111111
 19|1111111122222222222222222222222222222222222222222222222222222222111111
-20|1111111124444444444444444444444444444444444444444444444455555442111111
+20|1111111125555555555555555555555555555555555555555555555544444552111111
 21|1111111122222222222222222222222222222222222222222222222222222222111111
 22|1111111111111111111111111111111111111111111111111111111111111111111111
 23|;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
