+++ tests/screenshots/tests-picker_ui_snap.lua---narrow---query_main_top
@@ -1,0 +1,51 @@
+--|---------|---------|---------|---------|---------|---------|---------|
+01|                                                                      
+02|~                                                                     
+03|~       ╭ FFFiles ─────────────────────────────────────────────╮      
+04|~       │> main                                         19/32  │      
+05|~       ├──────────────────────────────────────────────────────┤      
+06|~       │  main.rs src                                         │      
+07|~       │  main_helper.rs src                                  │      
+08|~       │  main_loop.rs src                                    │      
+09|~       │  main_runner.rs src                                  │      
+10|~       │  main_utils.rs src                                   │      
+11|~       │  main_test.rs tests                                  │      
+12|~       │  intro.md docs                                       │      
+13|~       │  input.tsx src/components                            │      
+14|~       │  integration.rs tests                                │      
+15|~       │  contributing.md docs                                │      
+16|~       │  changelog.md docs                                   │      
+17|~       │  menu.tsx src/components                             │      
+18|~       │  regression.rs tests                                 │      
+19|~       ╰──────────────────────────────────────────────────────╯      
+20|~                                                                     
+21|~                                                                     
+22|~                                                                     
+23|[No Name]                                           0,1            All
+24|-- INSERT --                                        1,7           All 
+
+--|---------|---------|---------|---------|---------|---------|---------|
+01|0000000000000000000000000000000000000000000000000000000000000000000000
+02|1111111111111111111111111111111111111111111111111111111111111111111111
+03|1111111123333333332222222222222222222222222222222222222222222222111111
+04|1111111124444444444444444444444444444444444444444444444455555442111111
+05|1111111122222222222222222222222222222222222222222222222222222222111111
+06|1111111126677778888999888888888888888888888888888888888888888882111111
+07|11111111255777744444444444:::44444444444444444444444444444444442111111
+08|111111112557777444444444:::4444444444444444444444444444444444442111111
+09|11111111255777744444444444:::44444444444444444444444444444444442111111
+10|1111111125577774444444444:::444444444444444444444444444444444442111111
+11|111111112557777444444444:::::44444444444444444444444444444444442111111
+12|11111111255444444444::::4444444444444444444444444444444444444442111111
+13|111111112554444444444::::::::::::::44444444444444444444444444442111111
+14|11111111255444444444444444:::::444444444444444444444444444444442111111
+15|111111112554444444444444444::::444444444444444444444444444444442111111
+16|111111112554444444444444::::444444444444444444444444444444444442111111
+17|11111111255444444444::::::::::::::444444444444444444444444444442111111
+18|1111111125544444444444444:::::4444444444444444444444444444444442111111
+19|1111111122222222222222222222222222222222222222222222222222222222111111
+20|1111111111111111111111111111111111111111111111111111111111111111111111
+21|1111111111111111111111111111111111111111111111111111111111111111111111
+22|1111111111111111111111111111111111111111111111111111111111111111111111
+23|;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+24|<<<<<<<<<<<<==========================================================
