+++ tests/screenshots/tests-picker_ui_snap.lua---scrollbar---next_page_bottom
