+++ tests/screenshots/tests-picker_ui_snap.lua---wide---cursor_second_item_bottom
